Clean up TechnoExt
