+++ src/Ext/Techno/Body.cpp
@@ -2,21 +2,9 @@
 
 #include <BuildingClass.h>
 #include <HouseClass.h>
-#include <BulletClass.h>
-#include <BulletTypeClass.h>
 #include <ScenarioClass.h>
-#include <SpawnManagerClass.h>
-#include <InfantryClass.h>
-#include <ParticleSystemClass.h>
-#include <Unsorted.h>
-#include <BitFont.h>
 
-#include <Ext/Bullet/Body.h>
-#include <Ext/BulletType/Body.h>
-#include <Ext/WeaponType/Body.h>
 #include <Ext/House/Body.h>
-#include <Misc/FlyingStrings.h>
-#include <Utilities/EnumFunctions.h>
 
 template<> const DWORD Extension<TechnoClass>::Canary = 0x55555555;
 TechnoExt::ExtContainer TechnoExt::ExtMap;
@@ -33,420 +21,6 @@ TechnoExt::ExtData::~ExtData()
 	}
 }
 
-void TechnoExt::ExtData::ApplyInterceptor()
-{
-	auto const pThis = this->OwnerObject();
-	auto const pTypeExt = this->TypeExtData;
-
-	if (pTypeExt && pTypeExt->InterceptorType && !pThis->Target &&
-		!(pThis->WhatAmI() == AbstractType::Aircraft && pThis->GetHeight() <= 0))
-	{
-		BulletClass* pTargetBullet = nullptr;
-
-		for (auto const& [pBullet, pBulletExt] : BulletExt::ExtMap)
-		{
-			const auto pInterceptorType = pTypeExt->InterceptorType.get();
-			const auto& guardRange = pInterceptorType->GuardRange.Get(pThis);
-			const auto& minguardRange = pInterceptorType->MinimumGuardRange.Get(pThis);
-
-			auto distance = pBullet->Location.DistanceFrom(pThis->Location);
-
-			if (distance > guardRange || distance < minguardRange)
-				continue;
-
-			auto pBulletTypeExt = pBulletExt->TypeExtData;
-
-			if (!pBulletTypeExt || !pBulletTypeExt->Interceptable)
-				continue;
-
-			if (pBulletTypeExt->Armor.isset())
-			{
-				int weaponIndex = pThis->SelectWeapon(pBullet);
-				auto pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
-				double versus = GeneralUtils::GetWarheadVersusArmor(pWeapon->Warhead, pBulletTypeExt->Armor.Get());
-
-				if (versus == 0.0)
-					continue;
-			}
-
-			auto bulletOwner = pBullet->Owner ? pBullet->Owner->Owner : pBulletExt->FirerHouse;
-
-			if (EnumFunctions::CanTargetHouse(pInterceptorType->CanTargetHouses, pThis->Owner, bulletOwner))
-			{
-				pTargetBullet = pBullet;
-
-				if (pBulletExt->InterceptedStatus == InterceptedStatus::Targeted)
-					continue;
-
-				break;
-			}
-		}
-
-		if (pTargetBullet)
-			pThis->SetTarget(pTargetBullet);
-	}
-}
-
-// TODO : Merge into new AttachEffects
-bool TechnoExt::ExtData::CheckDeathConditions()
-{
-	auto const pTypeExt = this->TypeExtData;
-
-	if (!pTypeExt->AutoDeath_Behavior.isset())
-		return false;
-
-	auto const pThis = this->OwnerObject();
-	auto const pType = pThis->GetTechnoType();
-
-	// Self-destruction must be enabled
-	const auto howToDie = pTypeExt->AutoDeath_Behavior.Get();
-
-	// Death if no ammo
-	if (pType->Ammo > 0 && pThis->Ammo <= 0 && pTypeExt->AutoDeath_OnAmmoDepletion)
-	{
-		TechnoExt::KillSelf(pThis, howToDie);
-		return true;
-	}
-
-	// Death if countdown ends
-	if (pTypeExt->AutoDeath_AfterDelay > 0)
-	{
-		if (!this->AutoDeathTimer.HasStarted())
-		{
-			this->AutoDeathTimer.Start(pTypeExt->AutoDeath_AfterDelay);
-			HouseExt::ExtMap.Find(pThis->Owner)->OwnedTimedAutoDeathObjects.push_back(this);
-		}
-		else if (this->AutoDeathTimer.Completed())
-		{
-			TechnoExt::KillSelf(pThis, howToDie);
-			return true;
-		}
-
-	}
-	// TODO : Not working correctly, FIX THIS
-	auto existTechnoTypes = [pThis](const ValueableVector<TechnoTypeClass*>& vTypes, AffectedHouse affectedHouse, bool any, bool allowLimbo)
-	{
-		auto existSingleType = [pThis, affectedHouse, allowLimbo](const TechnoTypeClass* pType)
-		{
-			for (HouseClass* pHouse : *HouseClass::Array)
-			{
-				if (EnumFunctions::CanTargetHouse(affectedHouse, pThis->Owner, pHouse)
-					&& (allowLimbo ? pHouse->CountOwnedNow(pType) > 0 : pHouse->CountOwnedAndPresent(pType) > 0))
-					return true;
-			}
-
-			return false;
-		};
-
-		return any
-			? std::any_of(vTypes.begin(), vTypes.end(), existSingleType)
-			: std::all_of(vTypes.begin(), vTypes.end(), existSingleType);
-	};
-
-	// death if listed technos don't exist
-	if (!pTypeExt->AutoDeath_TechnosDontExist.empty())
-	{
-		if (!existTechnoTypes(pTypeExt->AutoDeath_TechnosDontExist, pTypeExt->AutoDeath_TechnosDontExist_Houses, !pTypeExt->AutoDeath_TechnosDontExist_Any, pTypeExt->AutoDeath_TechnosDontExist_AllowLimboed))
-		{
-			TechnoExt::KillSelf(pThis, howToDie);
-
-			return true;
-		}
-	}
-
-	// death if listed technos exist
-	if (!pTypeExt->AutoDeath_TechnosExist.empty())
-	{
-		if (existTechnoTypes(pTypeExt->AutoDeath_TechnosExist, pTypeExt->AutoDeath_TechnosExist_Houses, pTypeExt->AutoDeath_TechnosExist_Any, pTypeExt->AutoDeath_TechnosDontExist_AllowLimboed))
-		{
-			TechnoExt::KillSelf(pThis, howToDie);
-
-			return true;
-		}
-	}
-
-	return false;
-}
-
-void TechnoExt::ExtData::EatPassengers()
-{
-	auto const pThis = this->OwnerObject();
-	auto const pTypeExt = this->TypeExtData;
-
-	if (!TechnoExt::IsActive(pThis) || !pTypeExt->PassengerDeletionType)
-		return;
-
-	auto pDelType = pTypeExt->PassengerDeletionType.get();
-
-	if (pTypeExt && (pDelType->Rate > 0 || pDelType->UseCostAsRate))
-	{
-		if (pThis->Passengers.NumPassengers > 0)
-		{
-			// Passengers / CargoClass is essentially a stack, last in, first out (LIFO) kind of data structure
-			FootClass* pPassenger = nullptr;          // Passenger to potentially delete
-			FootClass* pPreviousPassenger = nullptr;  // Passenger immediately prior to the deleted one in the stack
-			ObjectClass* pLastPassenger = nullptr;    // Passenger that is last in the stack
-			auto pCurrentPassenger = pThis->Passengers.GetFirstPassenger();
-
-			// Find the first entered passenger that is eligible for deletion.
-			while (pCurrentPassenger)
-			{
-				if (EnumFunctions::CanTargetHouse(pDelType->AllowedHouses, pThis->Owner, pCurrentPassenger->Owner))
-				{
-					pPreviousPassenger = abstract_cast<FootClass*>(pLastPassenger);;
-					pPassenger = pCurrentPassenger;
-				}
-
-				pLastPassenger = pCurrentPassenger;
-				pCurrentPassenger = abstract_cast<FootClass*>(pCurrentPassenger->NextObject);
-			}
-
-			if (!pPassenger)
-			{
-				this->PassengerDeletionTimer.Stop();
-				return;
-			}
-
-			if (!this->PassengerDeletionTimer.IsTicking()) // Execute only if timer has been stopped or not started
-			{
-				int timerLength = 0;
-
-				if (pDelType->UseCostAsRate)
-				{
-					// Use passenger cost as countdown.
-					timerLength = (int)(pPassenger->GetTechnoType()->Cost * pDelType->CostMultiplier);
-
-					if (pDelType->CostRateCap.isset())
-						timerLength = std::min(timerLength, pDelType->CostRateCap.Get());
-				}
-				else
-				{
-					// Use explicit rate optionally multiplied by unit size as countdown.
-					timerLength = pDelType->Rate;
-
-					if (pDelType->Rate_SizeMultiply && pPassenger->GetTechnoType()->Size > 1.0)
-						timerLength *= (int)(pPassenger->GetTechnoType()->Size + 0.5);
-				}
-
-				this->PassengerDeletionTimer.Start(timerLength);
-			}
-			else if (this->PassengerDeletionTimer.Completed()) // Execute only if timer has ran out after being started
-			{
-				--pThis->Passengers.NumPassengers;
-
-				if (pLastPassenger)
-					pLastPassenger->NextObject = nullptr;
-
-				if (pPreviousPassenger)
-					pPreviousPassenger->NextObject = pPassenger->NextObject;
-
-				if (pThis->Passengers.NumPassengers <= 0)
-					pThis->Passengers.FirstPassenger = nullptr;
-
-				if (auto const pPassengerType = pPassenger->GetTechnoType())
-				{
-					if (pDelType->ReportSound.isset())
-						VocClass::PlayAt(pDelType->ReportSound.Get(), pThis->GetCoords(), nullptr);
-
-					if (pDelType->Anim.isset())
-					{
-						const auto pAnimType = pDelType->Anim.Get();
-						if (auto const pAnim = GameCreate<AnimClass>(pAnimType, pThis->Location))
-						{
-							pAnim->SetOwnerObject(pThis);
-							pAnim->Owner = pThis->Owner;
-						}
-					}
-
-					// Check if there is money refund
-					if (pDelType->Soylent &&
-						EnumFunctions::CanTargetHouse(pDelType->SoylentAllowedHouses, pThis->Owner, pPassenger->Owner))
-					{
-						int nMoneyToGive = (int)(pPassenger->GetTechnoType()->GetRefund(pPassenger->Owner, true) * pDelType->SoylentMultiplier);
-
-						if (nMoneyToGive > 0)
-						{
-							pThis->Owner->GiveMoney(nMoneyToGive);
-							if (pDelType->DisplaySoylent)
-							{
-								FlyingStrings::AddMoneyString(nMoneyToGive, pThis->Owner,
-									pDelType->DisplaySoylentToHouses, pThis->Location, pDelType->DisplaySoylentOffset);
-							}
-						}
-					}
-
-					// Handle gunner change.
-					if (pThis->GetTechnoType()->Gunner)
-					{
-						if (auto const pFoot = abstract_cast<FootClass*>(pThis))
-						{
-							pFoot->RemoveGunner(pPassenger);
-
-							if (pThis->Passengers.NumPassengers > 0)
-								pFoot->ReceiveGunner(pThis->Passengers.FirstPassenger);
-						}
-					}
-
-					auto pSource = pDelType->DontScore ? nullptr : pThis;
-					pPassenger->KillPassengers(pSource);
-					pPassenger->RegisterDestruction(pSource);
-					pPassenger->UnInit();
-				}
-
-				this->PassengerDeletionTimer.Stop();
-			}
-		}
-		else
-		{
-			this->PassengerDeletionTimer.Stop();
-		}
-	}
-}
-
-void TechnoExt::ExtData::UpdateShield()
-{
-	auto const pTypeExt = this->TypeExtData;
-
-	// Set current shield type if it is not set.
-	if (!this->CurrentShieldType->Strength && pTypeExt->ShieldType->Strength)
-		this->CurrentShieldType = pTypeExt->ShieldType;
-
-	// Create shield class instance if it does not exist.
-	if (this->CurrentShieldType && this->CurrentShieldType->Strength && !this->Shield)
-		this->Shield = std::make_unique<ShieldClass>(this->OwnerObject());
-
-	if (const auto pShieldData = this->Shield.get())
-		pShieldData->AI();
-}
-
-void TechnoExt::ExtData::UpdateOnTunnelEnter()
-{
-	if (!this->IsInTunnel)
-	{
-		if (const auto pShieldData = this->Shield.get())
-			pShieldData->SetAnimationVisibility(false);
-
-		for (auto& pLaserTrail : this->LaserTrails)
-		{
-			pLaserTrail->Visible = false;
-			pLaserTrail->LastLocation = { };
-		}
-
-		this->IsInTunnel = true;
-	}
-}
-
-void TechnoExt::ExtData::ApplySpawnLimitRange()
-{
-	auto const pThis = this->OwnerObject();
-	auto const pTypeExt = this->TypeExtData;
-
-	if (pTypeExt->Spawner_LimitRange)
-	{
-		if (auto const pManager = pThis->SpawnManager)
-		{
-			auto pTechnoType = pThis->GetTechnoType();
-			int weaponRange = 0;
-			int weaponRangeExtra = pTypeExt->Spawner_ExtraLimitRange * Unsorted::LeptonsPerCell;
-
-			auto setWeaponRange = [&weaponRange](WeaponTypeClass* pWeaponType)
-			{
-				if (pWeaponType && pWeaponType->Spawner && pWeaponType->Range > weaponRange)
-					weaponRange = pWeaponType->Range;
-			};
-
-			setWeaponRange(pTechnoType->Weapon[0].WeaponType);
-			setWeaponRange(pTechnoType->Weapon[1].WeaponType);
-			setWeaponRange(pTechnoType->EliteWeapon[0].WeaponType);
-			setWeaponRange(pTechnoType->EliteWeapon[1].WeaponType);
-
-			weaponRange += weaponRangeExtra;
-
-			if (pManager->Target && (pThis->DistanceFrom(pManager->Target) > weaponRange))
-				pManager->ResetTarget();
-		}
-	}
-}
-
-void TechnoExt::ExtData::UpdateTypeData(TechnoTypeClass* currentType)
-{
-	auto const pThis = this->OwnerObject();
-
-	if (this->LaserTrails.size())
-		this->LaserTrails.clear();
-
-	this->TypeExtData = TechnoTypeExt::ExtMap.Find(currentType);
-
-	// Recreate Laser Trails
-	for (auto const& entry : this->TypeExtData->LaserTrailData)
-	{
-		if (auto const pLaserType = LaserTrailTypeClass::Array[entry.idxType].get())
-		{
-			this->LaserTrails.push_back(std::make_unique<LaserTrailClass>(
-				pLaserType, pThis->Owner, entry.FLH, entry.IsOnTurret));
-		}
-	}
-
-	// Reset Shield
-	// This part should have been done by UpdateShield
-	// But that doesn't work correctly either, FIX THAT
-
-	// Reset AutoDeath Timer
-	if (this->AutoDeathTimer.HasStarted())
-	{
-		this->AutoDeathTimer.Stop();
-
-		auto hExt = HouseExt::ExtMap.Find(pThis->Owner);
-		auto it = std::find(hExt->OwnedTimedAutoDeathObjects.begin(), hExt->OwnedTimedAutoDeathObjects.end(), this);
-		if (it != hExt->OwnedTimedAutoDeathObjects.end())
-			hExt->OwnedTimedAutoDeathObjects.erase(it);
-	}
-
-	// Reset PassengerDeletion Timer - TODO : unchecked
-	if (this->PassengerDeletionTimer.IsTicking() && this->TypeExtData->PassengerDeletionType && this->TypeExtData->PassengerDeletionType->Rate <= 0)
-		this->PassengerDeletionTimer.Stop();
-}
-
-void TechnoExt::ExtData::UpdateLaserTrails()
-{
-	auto const pThis = this->OwnerObject();
-
-	// LaserTrails update routine is in TechnoClass::AI hook because TechnoClass::Draw
-	// doesn't run when the object is off-screen which leads to visual bugs - Kerbiter
-	for (auto const& trail : this->LaserTrails)
-	{
-		if (pThis->CloakState == CloakState::Cloaked && !trail->Type->CloakVisible)
-			continue;
-
-		if (!this->IsInTunnel)
-			trail->Visible = true;
-
-		CoordStruct trailLoc = TechnoExt::GetFLHAbsoluteCoords(pThis, trail->FLH, trail->IsOnTurret);
-		if (pThis->CloakState == CloakState::Uncloaking && !trail->Type->CloakVisible)
-			trail->LastLocation = trailLoc;
-		else
-			trail->Update(trailLoc);
-	}
-}
-
-void TechnoExt::ExtData::InitializeLaserTrails()
-{
-	if (this->LaserTrails.size())
-		return;
-
-	if (auto pTypeExt = this->TypeExtData)
-	{
-		for (auto const& entry : pTypeExt->LaserTrailData)
-		{
-			if (auto const pLaserType = LaserTrailTypeClass::Array[entry.idxType].get())
-			{
-				this->LaserTrails.push_back(std::make_unique<LaserTrailClass>(
-					pLaserType, this->OwnerObject()->Owner, entry.FLH, entry.IsOnTurret));
-			}
-		}
-	}
-}
-
 bool TechnoExt::IsActive(TechnoClass* pThis)
 {
 	return
@@ -457,52 +31,6 @@ bool TechnoExt::IsActive(TechnoClass* pThis)
 		pThis->IsAlive &&
 		pThis->Health > 0 &&
 		!pThis->InLimbo;
-}
-
-// TODO: FS-21 FIX THIS
-void TechnoExt::ObjectKilledBy(TechnoClass* pVictim, TechnoClass* pKiller)
-{
-	if (auto pVictimTechno = static_cast<TechnoClass*>(pVictim))
-	{
-		auto pVictimTechnoData = TechnoExt::ExtMap.Find(pVictim);
-
-		if (pVictimTechnoData && pKiller)
-		{
-			TechnoClass* pObjectKiller;
-
-			if ((pKiller->GetTechnoType()->Spawned || pKiller->GetTechnoType()->MissileSpawn) && pKiller->SpawnOwner)
-				pObjectKiller = pKiller->SpawnOwner;
-			else
-				pObjectKiller = pKiller;
-
-			if (pObjectKiller && pObjectKiller->BelongsToATeam())
-			{
-				auto pKillerTechnoData = TechnoExt::ExtMap.Find(pObjectKiller);
-				auto const pFootKiller = abstract_cast<FootClass*>(pObjectKiller);
-				auto const pFocus = abstract_cast<TechnoClass*>(pFootKiller->Team->Focus);
-				/*
-				Debug::Log("DEBUG: pObjectKiller -> [%s] [%s] registered a kill of the type [%s]\n",
-					pFootKiller->Team->Type->ID, pObjectKiller->get_ID(), pVictim->get_ID());
-				*/
-				pKillerTechnoData->LastKillWasTeamTarget = false;
-				if (pFocus == pVictim)
-					pKillerTechnoData->LastKillWasTeamTarget = true;
-			}
-		}
-	}
-}
-
-void TechnoExt::ApplyMindControlRangeLimit(TechnoClass* pThis)
-{
-	if (auto pCapturer = pThis->MindControlledBy)
-	{
-		auto pCapturerExt = TechnoTypeExt::ExtMap.Find(pCapturer->GetTechnoType());
-		if (pCapturerExt && pCapturerExt->MindControlRangeLimit.Get() > 0 &&
-			pThis->DistanceFrom(pCapturer) > pCapturerExt->MindControlRangeLimit.Get())
-		{
-			pCapturer->CaptureManager->FreeUnit(pThis);
-		}
-	}
 }
 
 bool TechnoExt::IsHarvesting(TechnoClass* pThis)
@@ -547,280 +75,6 @@ bool TechnoExt::HasAvailableDock(TechnoClass* pThis)
 	return false;
 }
 
-// reversed from 6F3D60
-CoordStruct TechnoExt::GetFLHAbsoluteCoords(TechnoClass* pThis, CoordStruct pCoord, bool isOnTurret)
-{
-	auto const pType = pThis->GetTechnoType();
-	auto const pFoot = abstract_cast<FootClass*>(pThis);
-	Matrix3D mtx;
-
-	// Step 1: get body transform matrix
-	if (pFoot && pFoot->Locomotor)
-		mtx = pFoot->Locomotor->Draw_Matrix(nullptr);
-	else // no locomotor means no rotation or transform of any kind (f.ex. buildings) - Kerbiter
-		mtx.MakeIdentity();
-
-	// Steps 2-3: turret offset and rotation
-	if (isOnTurret && pThis->HasTurret())
-	{
-		TechnoTypeExt::ApplyTurretOffset(pType, &mtx);
-
-		double turretRad = pThis->TurretFacing().GetRadian<32>();
-		double bodyRad = pThis->PrimaryFacing.Current().GetRadian<32>();
-		float angle = (float)(turretRad - bodyRad);
-
-		mtx.RotateZ(angle);
-	}
-
-	// Step 4: apply FLH offset
-	mtx.Translate((float)pCoord.X, (float)pCoord.Y, (float)pCoord.Z);
-
-	auto result = mtx * Vector3D<float>::Empty;
-
-	// Resulting coords are mirrored along X axis, so we mirror it back
-	result.Y *= -1;
-
-	// Step 5: apply as an offset to global object coords
-	CoordStruct location = pThis->GetCoords();
-	location += { (int)result.X, (int)result.Y, (int)result.Z };
-
-	return location;
-}
-
-CoordStruct TechnoExt::GetBurstFLH(TechnoClass* pThis, int weaponIndex, bool& FLHFound)
-{
-	FLHFound = false;
-	CoordStruct FLH = CoordStruct::Empty;
-
-	if (!pThis || weaponIndex < 0)
-		return FLH;
-
-	auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
-
-	auto pInf = abstract_cast<InfantryClass*>(pThis);
-	auto& pickedFLHs = pExt->WeaponBurstFLHs;
-
-	if (pThis->Veterancy.IsElite())
-	{
-		if (pInf && pInf->IsDeployed())
-			pickedFLHs = pExt->EliteDeployedWeaponBurstFLHs;
-		else if (pInf && pInf->Crawling)
-			pickedFLHs = pExt->EliteCrouchedWeaponBurstFLHs;
-		else
-			pickedFLHs = pExt->EliteWeaponBurstFLHs;
-	}
-	else
-	{
-		if (pInf && pInf->IsDeployed())
-			pickedFLHs = pExt->DeployedWeaponBurstFLHs;
-		else if (pInf && pInf->Crawling)
-			pickedFLHs = pExt->CrouchedWeaponBurstFLHs;
-	}
-
-	if ((int)pickedFLHs[weaponIndex].size() > pThis->CurrentBurstIndex)
-	{
-		FLHFound = true;
-		FLH = pickedFLHs[weaponIndex][pThis->CurrentBurstIndex];
-	}
-
-	return FLH;
-}
-
-CoordStruct TechnoExt::GetSimpleFLH(InfantryClass* pThis, int weaponIndex, bool& FLHFound)
-{
-	FLHFound = false;
-	CoordStruct FLH = CoordStruct::Empty;
-
-	if (!pThis || weaponIndex < 0)
-		return FLH;
-
-	if (auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
-	{
-		Nullable<CoordStruct> pickedFLH;
-
-		if (pThis->IsDeployed())
-		{
-			if (weaponIndex == 0)
-				pickedFLH = pTypeExt->DeployedPrimaryFireFLH;
-			else if (weaponIndex == 1)
-				pickedFLH = pTypeExt->DeployedSecondaryFireFLH;
-		}
-		else
-		{
-			if (pThis->Crawling)
-			{
-				if (weaponIndex == 0)
-					pickedFLH = pTypeExt->PronePrimaryFireFLH;
-				else if (weaponIndex == 1)
-					pickedFLH = pTypeExt->ProneSecondaryFireFLH;
-			}
-		}
-
-		if (pickedFLH.isset())
-		{
-			FLH = pickedFLH.Get();
-			FLHFound = true;
-		}
-	}
-
-	return FLH;
-}
-
-void TechnoExt::KillSelf(TechnoClass* pThis, AutoDeathBehavior deathOption)
-{
-	switch (deathOption)
-	{
-
-	case AutoDeathBehavior::Vanish:
-	{
-		pThis->KillPassengers(pThis);
-		pThis->vt_entry_3A0();
-		pThis->Limbo();
-		pThis->RegisterKill(pThis->Owner);
-		pThis->UnInit();
-
-		return;
-	}
-
-	case AutoDeathBehavior::Sell:
-	{
-		if (auto pBld = abstract_cast<BuildingClass*>(pThis))
-		{
-			if (pBld->HasBuildUp)
-			{
-				pBld->Sell(true);
-
-				return;
-			}
-		}
-		if (Phobos::Config::DevelopmentCommands)
-			Debug::Log("[Developer warning] AutoDeath: [%s] can't be sold, killing it instead\n", pThis->get_ID());
-	}
-
-	default: //must be AutoDeathBehavior::Kill
-		pThis->ReceiveDamage(&pThis->Health, 0, RulesClass::Instance->C4Warhead, nullptr, true, false, nullptr);
-		// Due to Ares, ignoreDefense=true will prevent passenger/crew/hijacker from escaping
-		return;
-	}
-}
-
-void TechnoExt::UpdateSharedAmmo(TechnoClass* pThis)
-{
-	if (!pThis)
-		return;
-
-	if (const auto pType = pThis->GetTechnoType())
-	{
-		if (pType->OpenTopped && pThis->Passengers.NumPassengers > 0)
-		{
-			if (const auto pExt = TechnoTypeExt::ExtMap.Find(pType))
-			{
-				if (pExt->Ammo_Shared && pType->Ammo > 0)
-				{
-					auto passenger = pThis->Passengers.FirstPassenger;
-					TechnoTypeClass* passengerType;
-
-					do
-					{
-						passengerType = passenger->GetTechnoType();
-						auto pPassengerExt = TechnoTypeExt::ExtMap.Find(passengerType);
-
-						if (pPassengerExt && pPassengerExt->Ammo_Shared)
-						{
-							if (pExt->Ammo_Shared_Group < 0 || pExt->Ammo_Shared_Group == pPassengerExt->Ammo_Shared_Group)
-							{
-								if (pThis->Ammo > 0 && (passenger->Ammo < passengerType->Ammo))
-								{
-									pThis->Ammo--;
-									passenger->Ammo++;
-								}
-							}
-						}
-
-						passenger = static_cast<FootClass*>(passenger->NextObject);
-					}
-					while (passenger);
-				}
-			}
-		}
-	}
-}
-
-void TechnoExt::ApplyGainedSelfHeal(TechnoClass* pThis)
-{
-	int healthDeficit = pThis->GetTechnoType()->Strength - pThis->Health;
-
-	if (pThis->Health && healthDeficit > 0)
-	{
-		if (auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
-		{
-			bool isBuilding = pThis->WhatAmI() == AbstractType::Building;
-			bool isOrganic = pThis->WhatAmI() == AbstractType::Infantry || pThis->WhatAmI() == AbstractType::Unit && pThis->GetTechnoType()->Organic;
-			auto defaultSelfHealType = isBuilding ? SelfHealGainType::None : isOrganic ? SelfHealGainType::Infantry : SelfHealGainType::Units;
-			auto selfHealType = pExt->SelfHealGainType.Get(defaultSelfHealType);
-
-			if (selfHealType == SelfHealGainType::None)
-				return;
-
-			bool applyHeal = false;
-			int amount = 0;
-
-			if (selfHealType == SelfHealGainType::Infantry)
-			{
-				int count = RulesExt::Global()->InfantryGainSelfHealCap.isset() ?
-					std::min(std::max(RulesExt::Global()->InfantryGainSelfHealCap.Get(), 1), pThis->Owner->InfantrySelfHeal) :
-					pThis->Owner->InfantrySelfHeal;
-
-				amount = RulesClass::Instance->SelfHealInfantryAmount * count;
-
-				if (!(Unsorted::CurrentFrame % RulesClass::Instance->SelfHealInfantryFrames) && amount)
-					applyHeal = true;
-			}
-			else
-			{
-				int count = RulesExt::Global()->UnitsGainSelfHealCap.isset() ?
-					std::min(std::max(RulesExt::Global()->UnitsGainSelfHealCap.Get(), 1), pThis->Owner->UnitsSelfHeal) :
-					pThis->Owner->UnitsSelfHeal;
-
-				amount = RulesClass::Instance->SelfHealUnitAmount * count;
-
-				if (!(Unsorted::CurrentFrame % RulesClass::Instance->SelfHealUnitFrames) && amount)
-					applyHeal = true;
-			}
-
-			if (applyHeal && amount)
-			{
-				if (amount >= healthDeficit)
-					amount = healthDeficit;
-
-				bool wasDamaged = pThis->GetHealthPercentage() <= RulesClass::Instance->ConditionYellow;
-
-				pThis->Health += amount;
-
-				if (wasDamaged && (pThis->GetHealthPercentage() > RulesClass::Instance->ConditionYellow
-					|| pThis->GetHeight() < -10))
-				{
-					if (auto const pBuilding = abstract_cast<BuildingClass*>(pThis))
-					{
-						pBuilding->UpdatePlacement(PlacementType::Redraw);
-						pBuilding->ToggleDamagedAnims(false);
-					}
-
-					if (pThis->WhatAmI() == AbstractType::Unit || pThis->WhatAmI() == AbstractType::Building)
-					{
-						auto dmgParticle = pThis->DamageParticleSystem;
-
-						if (dmgParticle)
-							dmgParticle->UnInit();
-					}
-				}
-			}
-		}
-	}
-
-	return;
-}
-
 void TechnoExt::SyncIronCurtainStatus(TechnoClass* pFrom, TechnoClass* pTo)
 {
 	if (pFrom->IsIronCurtained() && !pFrom->ForceShielded)
@@ -831,93 +85,6 @@ void TechnoExt::SyncIronCurtainStatus(TechnoClass* pFrom, TechnoClass* pTo)
 			pTo->IronCurtain(pFrom->IronCurtainTimer.GetTimeLeft(), pFrom->Owner, false);
 			pTo->IronTintStage = pFrom->IronTintStage;
 		}
-	}
-}
-
-void TechnoExt::DrawSelfHealPips(TechnoClass* pThis, Point2D* pLocation, RectangleStruct* pBounds)
-{
-	bool drawPip = false;
-	bool isInfantryHeal = false;
-	int selfHealFrames = 0;
-
-	if (auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
-	{
-		if (pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::None)
-			return;
-
-		bool hasInfantrySelfHeal = pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::Infantry;
-		bool hasUnitSelfHeal = pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::Units;
-		bool isOrganic = false;
-
-		if (pThis->WhatAmI() == AbstractType::Infantry ||
-			pThis->GetTechnoType()->Organic && pThis->WhatAmI() == AbstractType::Unit)
-		{
-			isOrganic = true;
-		}
-
-		if (pThis->Owner->InfantrySelfHeal > 0 && (hasInfantrySelfHeal || isOrganic))
-		{
-			drawPip = true;
-			selfHealFrames = RulesClass::Instance->SelfHealInfantryFrames;
-			isInfantryHeal = true;
-		}
-		else if (pThis->Owner->UnitsSelfHeal > 0 && (hasUnitSelfHeal || (pThis->WhatAmI() == AbstractType::Unit && !isOrganic)))
-		{
-			drawPip = true;
-			selfHealFrames = RulesClass::Instance->SelfHealUnitFrames;
-		}
-	}
-
-	if (drawPip)
-	{
-		Valueable<Point2D> pipFrames;
-		bool isSelfHealFrame = false;
-		int xOffset = 0;
-		int yOffset = 0;
-
-		if (Unsorted::CurrentFrame % selfHealFrames <= 5
-			&& pThis->Health < pThis->GetTechnoType()->Strength)
-		{
-			isSelfHealFrame = true;
-		}
-
-		if (pThis->WhatAmI() == AbstractType::Unit || pThis->WhatAmI() == AbstractType::Aircraft)
-		{
-			auto& offset = RulesExt::Global()->Pips_SelfHeal_Units_Offset.Get();
-			pipFrames = RulesExt::Global()->Pips_SelfHeal_Units;
-			xOffset = offset.X;
-			yOffset = offset.Y + pThis->GetTechnoType()->PixelSelectionBracketDelta;
-		}
-		else if (pThis->WhatAmI() == AbstractType::Infantry)
-		{
-			auto& offset = RulesExt::Global()->Pips_SelfHeal_Infantry_Offset.Get();
-			pipFrames = RulesExt::Global()->Pips_SelfHeal_Infantry;
-			xOffset = offset.X;
-			yOffset = offset.Y + pThis->GetTechnoType()->PixelSelectionBracketDelta;
-		}
-		else
-		{
-			auto pType = abstract_cast<BuildingTypeClass*>(pThis->GetTechnoType());
-			int fHeight = pType->GetFoundationHeight(false);
-			int yAdjust = -Unsorted::CellHeightInPixels / 2;
-
-			auto& offset = RulesExt::Global()->Pips_SelfHeal_Buildings_Offset.Get();
-			pipFrames = RulesExt::Global()->Pips_SelfHeal_Buildings;
-			xOffset = offset.X + Unsorted::CellWidthInPixels / 2 * fHeight;
-			yOffset = offset.Y + yAdjust * fHeight + pType->Height * yAdjust;
-		}
-
-		int pipFrame = isInfantryHeal ? pipFrames.Get().X : pipFrames.Get().Y;
-
-		Point2D position = { pLocation->X + xOffset, pLocation->Y + yOffset };
-
-		auto flags = BlitterFlags::bf_400 | BlitterFlags::Centered;
-
-		if (isSelfHealFrame)
-			flags = flags | BlitterFlags::Darken;
-
-		DSurface::Temp->DrawSHP(FileSystem::PALETTE_PAL, FileSystem::PIPS_SHP,
-		pipFrame, &position, pBounds, flags, 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
 	}
 }
 
@@ -934,77 +101,6 @@ double TechnoExt::GetCurrentSpeedMultiplier(FootClass* pThis)
 
 	return pThis->SpeedMultiplier * houseMultiplier *
 		(pThis->HasAbility(Ability::Faster) ? RulesClass::Instance->VeteranSpeed : 1.0);
-}
-
-void TechnoExt::UpdateMindControlAnim(TechnoClass* pThis)
-{
-	if (const auto pExt = TechnoExt::ExtMap.Find(pThis))
-	{
-		if (pThis->IsMindControlled())
-		{
-			if (pThis->MindControlRingAnim && !pExt->MindControlRingAnimType)
-			{
-				pExt->MindControlRingAnimType = pThis->MindControlRingAnim->Type;
-			}
-			else if (!pThis->MindControlRingAnim && pExt->MindControlRingAnimType &&
-				pThis->CloakState == CloakState::Uncloaked && !pThis->InLimbo && pThis->IsAlive)
-			{
-				auto coords = CoordStruct::Empty;
-				coords = *pThis->GetCoords(&coords);
-				int offset = 0;
-
-				if (const auto pBuilding = specific_cast<BuildingClass*>(pThis))
-					offset = Unsorted::LevelHeight * pBuilding->Type->Height;
-				else
-					offset = pThis->GetTechnoType()->MindControlRingOffset;
-
-				coords.Z += offset;
-				auto anim = GameCreate<AnimClass>(pExt->MindControlRingAnimType, coords, 0, 1);
-
-				if (anim)
-				{
-					pThis->MindControlRingAnim = anim;
-					pThis->MindControlRingAnim->SetOwnerObject(pThis);
-
-					if (pThis->WhatAmI() == AbstractType::Building)
-						pThis->MindControlRingAnim->ZAdjust = -1024;
-				}
-			}
-		}
-		else if (pExt->MindControlRingAnimType)
-		{
-			pExt->MindControlRingAnimType = nullptr;
-		}
-	}
-}
-
-void TechnoExt::DisplayDamageNumberString(TechnoClass* pThis, int damage, bool isShieldDamage)
-{
-	if (!pThis || damage == 0)
-		return;
-
-	const auto pExt = TechnoExt::ExtMap.Find(pThis);
-	ColorStruct color;
-
-	if (!isShieldDamage)
-		color = damage > 0 ? ColorStruct { 255, 0, 0 } : ColorStruct { 0, 255, 0 };
-	else
-		color = damage > 0 ? ColorStruct { 0, 160, 255 } : ColorStruct { 0, 255, 230 };
-
-	auto coords = pThis->GetRenderCoords();
-	int maxOffset = Unsorted::CellWidthInPixels / 2;
-	int width = 0, height = 0;
-	wchar_t damageStr[0x20];
-	swprintf_s(damageStr, L"%d", damage);
-
-	BitFont::Instance->GetTextDimension(damageStr, &width, &height, 120);
-
-	if (pExt->DamageNumberOffset >= maxOffset || pExt->DamageNumberOffset.empty())
-		pExt->DamageNumberOffset = -maxOffset;
-
-	FlyingStrings::Add(damageStr, coords, color, Point2D { pExt->DamageNumberOffset - (width / 2), 0 });
-
-	pExt->DamageNumberOffset = pExt->DamageNumberOffset + width;
 }
 
 CoordStruct TechnoExt::PassengerKickOutLocation(TechnoClass* pThis, FootClass* pPassenger, int maxAttempts = 1)
