+++ src/Ext/Techno/Ext.cpp
@@ -1,0 +1,272 @@
+#include "Body.h"
+
+#include <Misc/FlyingStrings.h>
+#include <BitFont.h>
+
+void TechnoExt::ExtData::InitializeLaserTrails()
+{
+	if (this->LaserTrails.size())
+		return;
+
+	if (auto pTypeExt = this->TypeExtData)
+	{
+		for (auto const& entry : pTypeExt->LaserTrailData)
+		{
+			if (auto const pLaserType = LaserTrailTypeClass::Array[entry.idxType].get())
+			{
+				this->LaserTrails.push_back(std::make_unique<LaserTrailClass>(
+					pLaserType, this->OwnerObject()->Owner, entry.FLH, entry.IsOnTurret));
+			}
+		}
+	}
+}
+
+void TechnoExt::ObjectKilledBy(TechnoClass* pVictim, TechnoClass* pKiller)
+{
+	TechnoClass* pObjectKiller = ((pKiller->GetTechnoType()->Spawned || pKiller->GetTechnoType()->MissileSpawn) && pKiller->SpawnOwner) ?
+		pKiller->SpawnOwner : pKiller;
+
+	if (pObjectKiller && pObjectKiller->BelongsToATeam())
+	{
+		if (auto const pFootKiller = generic_cast<FootClass*>(pObjectKiller))
+		{
+			auto pKillerTechnoData = TechnoExt::ExtMap.Find(pObjectKiller);
+			pKillerTechnoData->LastKillWasTeamTarget = pFootKiller->Team->Focus == pVictim;
+		}
+	}
+}
+
+// reversed from 6F3D60
+CoordStruct TechnoExt::GetFLHAbsoluteCoords(TechnoClass* pThis, CoordStruct pCoord, bool isOnTurret)
+{
+	auto const pType = pThis->GetTechnoType();
+	auto const pFoot = abstract_cast<FootClass*>(pThis);
+	Matrix3D mtx;
+
+	// Step 1: get body transform matrix
+	if (pFoot && pFoot->Locomotor)
+		mtx = pFoot->Locomotor->Draw_Matrix(nullptr);
+	else // no locomotor means no rotation or transform of any kind (f.ex. buildings) - Kerbiter
+		mtx.MakeIdentity();
+
+	// Steps 2-3: turret offset and rotation
+	if (isOnTurret && pThis->HasTurret())
+	{
+		TechnoTypeExt::ApplyTurretOffset(pType, &mtx);
+
+		double turretRad = pThis->TurretFacing().GetRadian<32>();
+		double bodyRad = pThis->PrimaryFacing.Current().GetRadian<32>();
+		float angle = (float)(turretRad - bodyRad);
+
+		mtx.RotateZ(angle);
+	}
+
+	// Step 4: apply FLH offset
+	mtx.Translate((float)pCoord.X, (float)pCoord.Y, (float)pCoord.Z);
+
+	auto result = mtx * Vector3D<float>::Empty;
+
+	// Resulting coords are mirrored along X axis, so we mirror it back
+	result.Y *= -1;
+
+	// Step 5: apply as an offset to global object coords
+	CoordStruct location = pThis->GetCoords();
+	location += { (int)result.X, (int)result.Y, (int)result.Z };
+
+	return location;
+}
+
+CoordStruct TechnoExt::GetBurstFLH(TechnoClass* pThis, int weaponIndex, bool& FLHFound)
+{
+	FLHFound = false;
+	CoordStruct FLH = CoordStruct::Empty;
+
+	if (!pThis || weaponIndex < 0)
+		return FLH;
+
+	auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+
+	auto pInf = abstract_cast<InfantryClass*>(pThis);
+	auto& pickedFLHs = pExt->WeaponBurstFLHs;
+
+	if (pThis->Veterancy.IsElite())
+	{
+		if (pInf && pInf->IsDeployed())
+			pickedFLHs = pExt->EliteDeployedWeaponBurstFLHs;
+		else if (pInf && pInf->Crawling)
+			pickedFLHs = pExt->EliteCrouchedWeaponBurstFLHs;
+		else
+			pickedFLHs = pExt->EliteWeaponBurstFLHs;
+	}
+	else
+	{
+		if (pInf && pInf->IsDeployed())
+			pickedFLHs = pExt->DeployedWeaponBurstFLHs;
+		else if (pInf && pInf->Crawling)
+			pickedFLHs = pExt->CrouchedWeaponBurstFLHs;
+	}
+
+	if ((int)pickedFLHs[weaponIndex].size() > pThis->CurrentBurstIndex)
+	{
+		FLHFound = true;
+		FLH = pickedFLHs[weaponIndex][pThis->CurrentBurstIndex];
+	}
+
+	return FLH;
+}
+
+CoordStruct TechnoExt::GetSimpleFLH(InfantryClass* pThis, int weaponIndex, bool& FLHFound)
+{
+	FLHFound = false;
+	CoordStruct FLH = CoordStruct::Empty;
+
+	if (!pThis || weaponIndex < 0)
+		return FLH;
+
+	if (auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
+	{
+		Nullable<CoordStruct> pickedFLH;
+
+		if (pThis->IsDeployed())
+		{
+			if (weaponIndex == 0)
+				pickedFLH = pTypeExt->DeployedPrimaryFireFLH;
+			else if (weaponIndex == 1)
+				pickedFLH = pTypeExt->DeployedSecondaryFireFLH;
+		}
+		else
+		{
+			if (pThis->Crawling)
+			{
+				if (weaponIndex == 0)
+					pickedFLH = pTypeExt->PronePrimaryFireFLH;
+				else if (weaponIndex == 1)
+					pickedFLH = pTypeExt->ProneSecondaryFireFLH;
+			}
+		}
+
+		if (pickedFLH.isset())
+		{
+			FLH = pickedFLH.Get();
+			FLHFound = true;
+		}
+	}
+
+	return FLH;
+}
+
+void TechnoExt::DisplayDamageNumberString(TechnoClass* pThis, int damage, bool isShieldDamage)
+{
+	if (!pThis || damage == 0)
+		return;
+
+	const auto pExt = TechnoExt::ExtMap.Find(pThis);
+	ColorStruct color;
+
+	if (!isShieldDamage)
+		color = damage > 0 ? ColorStruct { 255, 0, 0 } : ColorStruct { 0, 255, 0 };
+	else
+		color = damage > 0 ? ColorStruct { 0, 160, 255 } : ColorStruct { 0, 255, 230 };
+
+	auto coords = pThis->GetRenderCoords();
+	int maxOffset = Unsorted::CellWidthInPixels / 2;
+	int width = 0, height = 0;
+	wchar_t damageStr[0x20];
+	swprintf_s(damageStr, L"%d", damage);
+
+	BitFont::Instance->GetTextDimension(damageStr, &width, &height, 120);
+
+	if (pExt->DamageNumberOffset >= maxOffset || pExt->DamageNumberOffset.empty())
+		pExt->DamageNumberOffset = -maxOffset;
+
+	FlyingStrings::Add(damageStr, coords, color, Point2D { pExt->DamageNumberOffset - (width / 2), 0 });
+
+	pExt->DamageNumberOffset = pExt->DamageNumberOffset + width;
+}
+
+void TechnoExt::DrawSelfHealPips(TechnoClass* pThis, Point2D* pLocation, RectangleStruct* pBounds)
+{
+	bool drawPip = false;
+	bool isInfantryHeal = false;
+	int selfHealFrames = 0;
+
+	if (auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
+	{
+		if (pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::None)
+			return;
+
+		bool hasInfantrySelfHeal = pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::Infantry;
+		bool hasUnitSelfHeal = pExt->SelfHealGainType.isset() && pExt->SelfHealGainType.Get() == SelfHealGainType::Units;
+		bool isOrganic = false;
+
+		if (pThis->WhatAmI() == AbstractType::Infantry ||
+			pThis->GetTechnoType()->Organic && pThis->WhatAmI() == AbstractType::Unit)
+		{
+			isOrganic = true;
+		}
+
+		if (pThis->Owner->InfantrySelfHeal > 0 && (hasInfantrySelfHeal || isOrganic))
+		{
+			drawPip = true;
+			selfHealFrames = RulesClass::Instance->SelfHealInfantryFrames;
+			isInfantryHeal = true;
+		}
+		else if (pThis->Owner->UnitsSelfHeal > 0 && (hasUnitSelfHeal || (pThis->WhatAmI() == AbstractType::Unit && !isOrganic)))
+		{
+			drawPip = true;
+			selfHealFrames = RulesClass::Instance->SelfHealUnitFrames;
+		}
+	}
+
+	if (drawPip)
+	{
+		Valueable<Point2D> pipFrames;
+		bool isSelfHealFrame = false;
+		int xOffset = 0;
+		int yOffset = 0;
+
+		if (Unsorted::CurrentFrame % selfHealFrames <= 5
+			&& pThis->Health < pThis->GetTechnoType()->Strength)
+		{
+			isSelfHealFrame = true;
+		}
+
+		if (pThis->WhatAmI() == AbstractType::Unit || pThis->WhatAmI() == AbstractType::Aircraft)
+		{
+			auto& offset = RulesExt::Global()->Pips_SelfHeal_Units_Offset.Get();
+			pipFrames = RulesExt::Global()->Pips_SelfHeal_Units;
+			xOffset = offset.X;
+			yOffset = offset.Y + pThis->GetTechnoType()->PixelSelectionBracketDelta;
+		}
+		else if (pThis->WhatAmI() == AbstractType::Infantry)
+		{
+			auto& offset = RulesExt::Global()->Pips_SelfHeal_Infantry_Offset.Get();
+			pipFrames = RulesExt::Global()->Pips_SelfHeal_Infantry;
+			xOffset = offset.X;
+			yOffset = offset.Y + pThis->GetTechnoType()->PixelSelectionBracketDelta;
+		}
+		else
+		{
+			auto pType = abstract_cast<BuildingTypeClass*>(pThis->GetTechnoType());
+			int fHeight = pType->GetFoundationHeight(false);
+			int yAdjust = -Unsorted::CellHeightInPixels / 2;
+
+			auto& offset = RulesExt::Global()->Pips_SelfHeal_Buildings_Offset.Get();
+			pipFrames = RulesExt::Global()->Pips_SelfHeal_Buildings;
+			xOffset = offset.X + Unsorted::CellWidthInPixels / 2 * fHeight;
+			yOffset = offset.Y + yAdjust * fHeight + pType->Height * yAdjust;
+		}
+
+		int pipFrame = isInfantryHeal ? pipFrames.Get().X : pipFrames.Get().Y;
+
+		Point2D position = { pLocation->X + xOffset, pLocation->Y + yOffset };
+
+		auto flags = BlitterFlags::bf_400 | BlitterFlags::Centered;
+
+		if (isSelfHealFrame)
+			flags = flags | BlitterFlags::Darken;
+
+		DSurface::Temp->DrawSHP(FileSystem::PALETTE_PAL, FileSystem::PIPS_SHP,
+		pipFrame, &position, pBounds, flags, 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
+	}
+}
