+++ src/Ext/Techno/Hooks.Firing.cpp
@@ -8,8 +8,7 @@
 #include <Ext/WeaponType/Body.h>
 #include <Utilities/EnumFunctions.h>
 
-// Weapon Selection
-
+#pragma region TechnoClass_SelectWeapon
 DEFINE_HOOK(0x6F3339, TechnoClass_WhatWeaponShouldIUse_Interceptor, 0x8)
 {
 	enum { SkipGameCode = 0x6F3341, ReturnValue = 0x6F3406 };
@@ -231,8 +230,9 @@ DEFINE_HOOK(0x5218F3, InfantryClass_WhatWeaponShouldIUse_DeployFireWeapon, 0x6)
 	return 0;
 }
 
-// Pre-Firing Checks
+#pragma endregion
 
+#pragma region TechnoClass_GetFireError
 DEFINE_HOOK(0x6FC339, TechnoClass_CanFire, 0x6)
 {
 	enum { CannotFire = 0x6FCB7E };
@@ -343,8 +343,9 @@ DEFINE_HOOK(0x6FC689, TechnoClass_CanFire_LandNavalTarget, 0x6)
 	return 0;
 }
 
-// Weapon Firing
+#pragma endregion
 
+#pragma region TechnoClass_Fire
 DEFINE_HOOK(0x6FE43B, TechnoClass_FireAt_OpenToppedDmgMult, 0x8)
 {
 	enum { ApplyDamageMult = 0x6FE45A, ContinueCheck = 0x6FE460 };
@@ -443,7 +444,6 @@ DEFINE_HOOK(0x6FF43F, TechnoClass_FireAt_FeedbackWeapon, 0x6)
 	return 0;
 }
 
-
 DEFINE_HOOK(0x6FF660, TechnoClass_FireAt_Interceptor, 0x6)
 {
 	GET(TechnoClass* const, pSource, ESI);
@@ -496,8 +496,262 @@ DEFINE_HOOK(0x6FF4CC, TechnoClass_FireAt_ToggleLaserWeaponIndex, 0x6)
 	return 0;
 }
 
+#pragma endregion
+
+#pragma region TechnoClass_GetFLH
 // Feature: Allow Units using AlternateFLHs - by Trsdy
 // I don't want to rewrite something new, so I use the Infantry one directly
 // afaik it has no check for infantry-specific stuff here so far
 // and neither Ares nor Phobos has touched it, even that crawling flh one was in TechnoClass
 DEFINE_JUMP(VTABLE, 0x7F5D20, 0x523250);// Redirect UnitClass::GetFLH to InfantryClass::GetFLH (used to be TechnoClass::GetFLH)
+
+DEFINE_HOOK(0x6F3B37, TechnoClass_GetFLH_BurstFLH_1, 0x7)
+{
+	GET(TechnoClass*, pThis, EBX);
+	GET_STACK(int, weaponIndex, STACK_OFFSET(0xD8, 0x8));
+	bool FLHFound = false;
+	CoordStruct FLH = CoordStruct::Empty;
+
+	FLH = TechnoExt::GetBurstFLH(pThis, weaponIndex, FLHFound);
+
+	if (!FLHFound)
+	{
+		if (auto pInf = abstract_cast<InfantryClass*>(pThis))
+			FLH = TechnoExt::GetSimpleFLH(pInf, weaponIndex, FLHFound);
+	}
+
+	if (FLHFound)
+	{
+		R->ECX(FLH.X);
+		R->EBP(FLH.Y);
+		R->EAX(FLH.Z);
+	}
+
+	return 0;
+}
+
+DEFINE_HOOK(0x6F3C88, TechnoClass_GetFLH_BurstFLH_2, 0x6)
+{
+	GET(TechnoClass*, pThis, EBX);
+	GET_STACK(int, weaponIndex, STACK_OFFSET(0xD8, 0x8));
+	bool FLHFound = false;
+
+	TechnoExt::GetBurstFLH(pThis, weaponIndex, FLHFound);
+
+	if (FLHFound)
+		R->EAX(0);
+
+	return 0;
+}
+#pragma endregion
+
+// Reimplements the game function with few changes / optimizations
+DEFINE_HOOK(0x7012C2, TechnoClass_WeaponRange, 0x8)
+{
+	enum { ReturnResult = 0x70138F };
+
+	GET(TechnoClass*, pThis, ECX);
+	GET_STACK(int, weaponIndex, STACK_OFFSET(0x8, 0x4));
+
+	int result = 0;
+	auto pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
+
+	if (pWeapon)
+	{
+		result = pWeapon->Range;
+		auto pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+
+		if (pThis->GetTechnoType()->OpenTopped && !pTypeExt->OpenTopped_IgnoreRangefinding)
+		{
+			int smallestRange = INT32_MAX;
+			auto pPassenger = pThis->Passengers.FirstPassenger;
+
+			while (pPassenger && (pPassenger->AbstractFlags & AbstractFlags::Foot) != AbstractFlags::None)
+			{
+				int openTWeaponIndex = pPassenger->GetTechnoType()->OpenTransportWeapon;
+				int tWeaponIndex = 0;
+
+				if (openTWeaponIndex != -1)
+					tWeaponIndex = openTWeaponIndex;
+				else
+					tWeaponIndex = pPassenger->SelectWeapon(pThis->Target);
+
+				WeaponTypeClass* pTWeapon = pPassenger->GetWeapon(tWeaponIndex)->WeaponType;
+
+				if (pTWeapon && pTWeapon->FireInTransport)
+				{
+					if (pTWeapon->Range < smallestRange)
+						smallestRange = pTWeapon->Range;
+				}
+
+				pPassenger = abstract_cast<FootClass*>(pPassenger->NextObject);
+			}
+
+			if (result > smallestRange)
+				result = smallestRange;
+		}
+	}
+
+	R->EBX(result);
+	return ReturnResult;
+}
+
+// Basically a hack to make game and Ares pick laser properties from non-Primary weapons.
+DEFINE_HOOK(0x70E1A5, TechnoClass_GetTurretWeapon_LaserWeapon, 0x6)
+{
+	enum { ReturnResult = 0x70E1C7, Continue = 0x70E1AB };
+
+	GET(TechnoClass* const, pThis, ESI);
+
+	if (pThis->WhatAmI() == AbstractType::Building)
+	{
+		if (auto const pExt = TechnoExt::ExtMap.Find(pThis))
+		{
+			if (!pExt->CurrentLaserWeaponIndex.empty())
+			{
+				auto weaponStruct = pThis->GetWeapon(pExt->CurrentLaserWeaponIndex.get());
+				R->EAX(weaponStruct);
+				return ReturnResult;
+			}
+		}
+	}
+
+	// Restore overridden instructions.
+	R->EAX(pThis->GetTechnoType());
+	return Continue;
+}
+
+DEFINE_HOOK(0x6FD0B5, TechnoClass_RearmDelay_RandomDelay, 0x6)
+{
+	GET(WeaponTypeClass*, pWeapon, EDI);
+
+	const auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	auto range = pWeaponExt->ROF_RandomDelay.Get(RulesExt::Global()->ROF_RandomDelay);
+
+	R->EAX(GeneralUtils::GetRangedRandomOrSingleValue(range));
+	return 0;
+}
+
+DEFINE_HOOK(0x6FD054, TechnoClass_RearmDelay_ForceFullDelay, 0x6)
+{
+	enum { ApplyFullRearmDelay = 0x6FD09E };
+
+	GET(TechnoClass*, pThis, ESI);
+
+	// Currently only used with infantry, so a performance saving measure.
+	if (pThis->WhatAmI() == AbstractType::Infantry)
+	{
+		if (const auto pExt = TechnoExt::ExtMap.Find(pThis))
+		{
+			if (pExt->ForceFullRearmDelay)
+			{
+				pExt->ForceFullRearmDelay = false;
+				pThis->CurrentBurstIndex = 0;
+				return ApplyFullRearmDelay;
+			}
+		}
+	}
+
+	return 0;
+}
+
+// Issue #271: Separate burst delay for weapon type
+// Author: Starkku
+DEFINE_HOOK(0x6FD05E, TechnoClass_RearmDelay_BurstDelays, 0x7)
+{
+	GET(TechnoClass*, pThis, ESI);
+	GET(WeaponTypeClass*, pWeapon, EDI);
+
+	const auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	int burstDelay = pWeaponExt->GetBurstDelay(pThis->CurrentBurstIndex);
+
+	if (burstDelay >= 0)
+	{
+		R->EAX(burstDelay);
+		return 0x6FD099;
+	}
+
+	// Restore overridden instructions
+	GET(int, idxCurrentBurst, ECX);
+	return idxCurrentBurst <= 0 || idxCurrentBurst > 4 ? 0x6FD084 : 0x6FD067;
+}
+
+// Update ammo rounds
+DEFINE_HOOK(0x6FB086, TechnoClass_Reload_ReloadAmount, 0x8)
+{
+	GET(TechnoClass* const, pThis, ECX);
+
+	TechnoExt::UpdateSharedAmmo(pThis);
+
+	return 0;
+}
+
+DEFINE_HOOK(0x5209A7, InfantryClass_FiringAI_BurstDelays, 0x8)
+{
+	enum { Continue = 0x5209CD, ReturnFromFunction = 0x520AD9 };
+
+	GET(InfantryClass*, pThis, EBP);
+	GET(int, firingFrame, EDX);
+
+	int weaponIndex = pThis->SelectWeapon(pThis->Target);
+	const auto pWeapon = pThis->GetWeapon(weaponIndex)->WeaponType;
+
+	if (!pWeapon)
+		return ReturnFromFunction;
+
+	const auto pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
+	int cumulativeDelay = 0;
+	int projectedDelay = 0;
+
+	// Calculate cumulative burst delay as well cumulative delay after next shot (projected delay).
+	if (pWeaponExt->Burst_FireWithinSequence)
+	{
+		for (int i = 0; i <= pThis->CurrentBurstIndex; i++)
+		{
+			int burstDelay = pWeaponExt->GetBurstDelay(i);
+			int delay = 0;
+
+			if (burstDelay > -1)
+				delay = burstDelay;
+			else
+				delay = ScenarioClass::Instance->Random.RandomRanged(3, 5);
+
+			// Other than initial delay, treat 0 frame delays as 1 frame delay due to per-frame processing.
+			if (i != 0)
+				delay = Math::max(delay, 1);
+
+			cumulativeDelay += delay;
+
+			if (i == pThis->CurrentBurstIndex)
+				projectedDelay = cumulativeDelay + delay;
+		}
+	}
+
+	if (pThis->IsFiring && pThis->Animation.Value == firingFrame + cumulativeDelay)
+	{
+		if (pWeaponExt->Burst_FireWithinSequence)
+		{
+			int frameCount = pThis->Type->Sequence->GetSequence(pThis->SequenceAnim).CountFrames;
+
+			// If projected frame for firing next shot goes beyond the sequence frame count, cease firing after this shot and start rearm timer.
+			if (firingFrame + projectedDelay > frameCount)
+			{
+				const auto pExt = TechnoExt::ExtMap.Find(pThis);
+				pExt->ForceFullRearmDelay = true;
+			}
+		}
+
+		R->EAX(weaponIndex); // Reuse the weapon index to save some time.
+		return Continue;
+	}
+
+	return ReturnFromFunction;
+}
+
+// Author: Otamaa
+DEFINE_HOOK(0x5223B3, InfantryClass_Approach_Target_DeployFireWeapon, 0x6)
+{
+	GET(InfantryClass*, pThis, ESI);
+	R->EDI(pThis->Type->DeployFireWeapon == -1 ? pThis->SelectWeapon(pThis->Target) : pThis->Type->DeployFireWeapon);
+	return 0x5223B9;
+}
