+++ src/Ext/Techno/Hooks.Transport.cpp
@@ -46,11 +46,11 @@ DEFINE_HOOK(0x701881, TechnoClass_ChangeHouse_Passenger_SyncOwner, 0x5)
 {
 	GET(TechnoClass*, pThis, ESI);
 
-	if (auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
+	if (FootClass* pPassenger = pThis->Passengers.GetFirstPassenger())
 	{
-		if (pTypeExt->Passengers_SyncOwner && pThis->Passengers.NumPassengers > 0)
+		if (auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
 		{
-			if (FootClass* pPassenger = pThis->Passengers.GetFirstPassenger())
+			if (pTypeExt->Passengers_SyncOwner && pThis->Passengers.NumPassengers > 0)
 			{
 				pPassenger->SetOwningHouse(pThis->Owner, false);
 
@@ -131,6 +131,40 @@ DEFINE_HOOK(0x737F80, TechnoClass_ReceiveDamage_Cargo_SyncOwner, 0x6)
 				if (pExt->OriginalPassengerOwner)
 					pPassenger->SetOwningHouse(pExt->OriginalPassengerOwner, false);
 			}
+		}
+	}
+
+	return 0;
+}
+
+// Customizable OpenTopped Properties
+// Author: Otamaa
+DEFINE_HOOK(0x6F72D2, TechnoClass_IsCloseEnoughToTarget_OpenTopped_RangeBonus, 0xC)
+{
+	GET(TechnoClass* const, pThis, ESI);
+
+	if (auto pTransport = pThis->Transporter)
+	{
+		if (auto pExt = TechnoTypeExt::ExtMap.Find(pTransport->GetTechnoType()))
+		{
+			R->EAX(pExt->OpenTopped_RangeBonus.Get(RulesClass::Instance->OpenToppedRangeBonus));
+			return 0x6F72DE;
+		}
+	}
+
+	return 0;
+}
+
+DEFINE_HOOK(0x71A82C, TemporalClass_AI_Opentopped_WarpDistance, 0xC)
+{
+	GET(TemporalClass* const, pThis, ESI);
+
+	if (auto pTransport = pThis->Owner->Transporter)
+	{
+		if (auto pExt = TechnoTypeExt::ExtMap.Find(pTransport->GetTechnoType()))
+		{
+			R->EDX(pExt->OpenTopped_WarpDistance.Get(RulesClass::Instance->OpenToppedWarpDistance));
+			return 0x71A838;
 		}
 	}
